@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.1"
+    compileSdkVersion 30
+    buildToolsVersion "29.0.3"
 
     defaultConfig {
         applicationId "com.peppe.ftpclient.androidftp"
-        minSdkVersion 11
-        targetSdkVersion 23
+        minSdkVersion 15
+        targetSdkVersion 30
         versionCode 1
         versionName "1.0"
     }
@@ -20,14 +20,14 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    testCompile 'junit:junit:4.12'
-    compile 'com.android.support:appcompat-v7:23.0.1'
-    compile 'com.android.support:recyclerview-v7:+'
-    compile 'com.android.support:design:23.0.1'
-    compile 'commons-net:commons-net:3.3'
-    compile 'org.solovyev.android.views:linear-layout-manager:0.5@aar'
-    compile 'de.greenrobot:eventbus:2.4.0'
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    testImplementation 'junit:junit:4.13'
+    implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation 'androidx.recyclerview:recyclerview:1.1.0'
+    implementation 'com.google.android.material:material:1.2.0'
+    implementation 'commons-net:commons-net:3.5'
+    implementation 'org.solovyev.android.views:linear-layout-manager:0.5@aar'
+    implementation 'de.greenrobot:eventbus:2.4.0'
     //compile 'commons-validator:commons-validator:1.4.1'
 }
 
