@@ -2,13 +2,13 @@ package com.peppe.ftpclient.androidftp.FTPFilesExplorer.FTPLocalExplorer;
 
 import android.content.Intent;
 import android.net.Uri;
-import android.support.v4.app.DialogFragment;
-import android.content.Context;
+import androidx.fragment.app.DialogFragment;
+
 import android.os.Bundle;
 import android.os.Environment;
-import android.support.annotation.Nullable;
-import android.support.v4.app.FragmentManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.FragmentManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.util.Log;
 import android.view.ActionMode;
 import android.view.Gravity;
@@ -18,19 +18,13 @@ import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
-import android.view.inputmethod.InputMethodManager;
 import android.webkit.MimeTypeMap;
-import android.widget.EditText;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.peppe.ftpclient.androidftp.FTPClientMain.MainActivity;
 import com.peppe.ftpclient.androidftp.FTPFilesExplorer.FTPBusEvents.UploadFilesEvent;
-import com.peppe.ftpclient.androidftp.FTPFilesExplorer.FilesAdapter;
 import com.peppe.ftpclient.androidftp.FTPFilesExplorer.FilesFragment;
 import com.peppe.ftpclient.androidftp.R;
-
-import org.apache.commons.net.ftp.FTPFile;
 
 import java.io.File;
 import java.io.FileInputStream;
